pyproject.toml
[project]
name = "test-bot"
version = "0.1.0"
description = "Discord Bot with FastAPI Dashboard"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "pycryptodome",
    "requests",
    "openai",
    "anthropic",
]

